requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
boto3==1.42.0
botocore==1.41.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
deepdiff==8.6.1
envs==1.4
frozenlist==1.8.0
idna==3.11
jmespath==1.0.1
multidict==6.7.0
orderly-set==5.5.0
propcache==0.4.1
pydantic==2.10.6
pycognito==2024.5.1
pycparser==2.23
PyJWT==2.10.1
pylitterbot==2024.2.7
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
typing_extensions==4.15.0
urllib3==1.26.20
yarl==1.22.0
